+++ src/Sentry.Unity.Editor/Android/AndroidManifestConfiguration.cs
@@ -218,9 +218,9 @@ public class AndroidManifestConfiguration
         androidManifest.SetAutoTraceIdGeneration(false);
         androidManifest.SetAutoSessionTracking(false);
         androidManifest.SetAutoAppLifecycleBreadcrumbs(false);
-        androidManifest.SetAnr(_options.AndroidAnrV2Enabled);
-        androidManifest.SetPersistentScopeObserver(_options.AndroidAnrV2Enabled);
-        androidManifest.SetAttachAnrThreadDump(_options.AndroidAnrV2Enabled && _options.AndroidAttachAnrThreadDump);
+        androidManifest.SetAnr(_options.AndroidNativeAnrEnabled);
+        androidManifest.SetPersistentScopeObserver(_options.AndroidNativeAnrEnabled);
+        androidManifest.SetAttachAnrThreadDump(_options.AndroidNativeAnrEnabled && _options.AndroidAttachAnrThreadDump);
         // Disable user interaction tracking to prevent conflicts with VR platforms (e.g., Oculus InputHooks)
         androidManifest.SetEnableUserInteractionBreadcrumbs(false);
         androidManifest.SetEnableUserInteractionTracing(false);